Move tests to topbar file
